% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepartureTime.R
\name{DepartureTime}
\alias{DepartureTime}
\title{Generate Departure times}
\usage{
DepartureTime(
  method = "H",
  dy = format(Sys.Date(), "\%Y"),
  dm = format(Sys.Date(), "\%m"),
  dd = format(Sys.Date(), "\%d"),
  tmin = 0,
  tmax = 24,
  res = 5,
  MMDD = TRUE,
  ptw = FALSE
)
}
\arguments{
\item{method}{(type: character) define sampling method. Optional values:

+ \code{`R`} OR \code{`Random`}: Simple random sampling;

+ \code{`S`} OR \code{`Systematic`}: Systematic sampling;

+ \code{`H`} OR \code{`Hybrid`}: Hybrid sampling;

+ \code{`W`} OR \code{`Walk`}: Constrained random walk sampling;

default: \strong{\code{"H"}}}

\item{dy}{year of analysis (format: YYYY)}

\item{dm}{month of analysis (format: MM)}

\item{dd}{day of analysis (format: DD)

default: \strong{default: system date}.}

\item{tmin}{lower limit of the time window}

\item{tmax}{upper limit of the time window

\strong{default: all day} (i.e 00:00 - 24:00)}

\item{res}{temporal resolution

\strong{default: 5 minutes}}

\item{MMDD}{date format of the output (TRUE / FALSE)

+ \code{TRUE} = MM/DD/YYYY

+ \code{FALSE} = DD/MM/YYYY

\strong{default: TRUE}}

\item{ptw}{Bolean: whether or not print limits of subsetted time-windows;

\strong{default: FALSE;}}
}
\value{
data.frame with departure times in rows
}
\description{
\code{DepartureTimes} generate data.frame with departure times for accessibility analyses with temporal resolution (e.g GTFS)
}
\examples{
\dontrun{
DepartureTime()

DepartureTime(method = "S",       # systematic sampling method
 dm = 5, dd = 15,                 # user-defined date: 15th May, 2019 (current year)
 tmin = 7, tmax = 10,             # user-defined time window (07:00 - 10:00)
 res = 15)                        # user-defined temporal resolution (15 minutes)
 }


}
